+++ src/ankit/assignment_1_5/Car.java
@@ -5,7 +5,7 @@
   	d. displaySpecification   
   Hint: One String, one boolean and one double variable needed. 
 */
-package ankit;
+package ankit.assignment_1_5;
 class Car{
 	
 	String companyName ;
